Đổi node database, fix chính sửa số lượng
- Hiển thị sản phẩm ở đúng node trong database
- Có thể chỉnh sửa số lượng sản phẩm muốn thêm (ràng buộc bởi số lượng khả dụng trong database)
- Public bet ở Main Activity để reuse

+++ app/src/main/java/vlu/mobileproject/modle/Products.java
@@ -117,13 +117,13 @@ public class Products implements Serializable {
         }
         return 0.0;
     }
-    public double getQuantityForMemory(String memory) {
+    public int getQuantityForMemory(String memory) {
         for (MemoryOption option : product_memoryOptions.values()) {
             if (option.getMemory() == memory) {
                 return option.getQuantity();
             }
         }
-        return 0.0;
+        return 0;
     }
 
     public String[] getMemory(){
